refactor(temper-server): simplify large server modules
